requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
Django==4.1.5
django-bootstrap-v5==1.0.11
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
